move framebuffer into opengl backend
