+++ renderlib/gfxOpenGL/CMakeLists.txt
@@ -10,6 +10,8 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/Backend.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Device.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Device.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Framebuffer.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Framebuffer.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/FSQ.cpp"
